Move deprecated function to there own file.
